requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
tortoise-orm>=0.24.0
pydantic>=2.0.0
aerich>=0.8.2 